Move strstore() function to inifile.cc
    dpkg-shlibdeps: warning: symbol _Z8strstorePKc used by
    debian/machinekit/usr/lib/libpyplugin.so.0 found in none of the
    libraries

Placed in `inifile.cc`, all `strstore()` using code shares a single
`stringtable` in `liblinuxcncini.so.0`.

`strstore()` is usually used directly with `.ini` file values, or else
with computed strings resulting from those values. In cases where this
isn't so clear, the sources are still in a binary object ultimately
linked against `liblinuxcncini.so.0`.

Signed-off-by: Mick <arceye@mgware.co.uk>